Add files via upload
All in one hack trainer that I made. I've used some resources from @neetjn. 90% of the things in this hack were found on my own. I may update it in the future, but for now, it is outdated. If you have the knowledge and tools required to fix it, go for it. I am not offering to help anyone with this.